# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LucaPOS API",
        "contact": {
            "email": "developers@tulivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "6d5b72038ed4abf362868c76845f5d14",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "400": {
                        "description": "Error no controlado."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/changepassword": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Cambiar la contraseña",
                "operationId": "d2a6cecf3f93b163a99cb4bcd71cf525",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se actualizó la contraseña",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "statusCode",
                                        "statusMessage",
                                        "result",
                                        "isSucess",
                                        "id",
                                        "size",
                                        "date",
                                        "error"
                                    ],
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "statusMessage": {
                                            "type": "string"
                                        },
                                        "result": {
                                            "$ref": "#/components/schemas/ErrorStatus"
                                        },
                                        "isSucess": {
                                            "type": "boolean"
                                        },
                                        "id": {
                                            "type": "integer"
                                        },
                                        "size": {
                                            "type": "integer"
                                        },
                                        "date": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se pudo actualizar la contraseña"
                    },
                    "401": {
                        "description": "Error no controlado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorStatus": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Inicie sesión con correo electrónico y contraseña para obtener el token de autenticación",
                "name": "Basado en token: bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}